src/main/java/com/syswin/temail/gateway/service/JSONStrUtil.java
/*
 * MIT License
 *
 * Copyright (c) 2019 Syswin
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package com.syswin.temail.gateway.service;

import java.util.regex.Pattern;
import org.apache.commons.lang3.StringUtils;

public class JSONStrUtil {

  private static final Pattern PATTERN = Pattern.compile("^\\{.*\\}$", Pattern.DOTALL);

  public static boolean isJSONStr(String str){
    if(StringUtils.isBlank(str)){
      return false;
    }
    return PATTERN.matcher(str).matches();
  }

  public static boolean isJSONBytes(byte[] bytes){
    if(bytes == null || bytes.length < 2){
      return false;
    }
    return bytes[0] == '{' && bytes[bytes.length-1] == '}';
  }

}
